Summary
This project is an incubator control system that uses a Raspberry Pi to control the temperature and humidity of an incubator. The system uses a DHT22 sensor to measure the temperature and humidity and a set of relays to control a heater, cooler, humidifier, and dehumidifier. The system also uses a Flask web application to display the current temperature, humidity, and status of the heater, cooler, humidifier, and dehumidifier on a web page. Additionally, it can store data locally in a .csv file, allows the user to set the start day of the incubation, automate the humidity and temperature changes, and provide a data visualization feature to show the data stored in the csv file on the web page.